pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = [
    {name = "Denilson"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref (==3.8.1)",
    "django (==5.1.6)",
    "django-filter (==25.1)",
    "djangorestframework (==3.15.2)",
    "markdown (==3.7)",
    "sqlparse (==0.5.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
